Changed int to size_t

+++ 16-binary_tree_is_perfect.c
@@ -27,7 +27,7 @@ size_t binary_tree_height(const binary_tree_t *tree)
  */
 int binary_tree_is_perfect(const binary_tree_t *tree)
 {
-	int ileft, iright;
+	size_t ileft, iright;
 
 	if (tree->left == NULL && tree->right == NULL)
 		return (1);
